Update about.css
Enhanced the "about us.html" page to better reflect our blogging platform's mission and values. The updated content now highlights our belief in the power of sharing diverse stories and experiences, aiming to foster a vibrant community of writers and readers. It includes details on our mission to empower individuals to freely express themselves and connect with a global audience. Visit the About Us section to explore how we support creativity and meaningful engagement on our platform.